Removed business logic from user endpoints and move it to handlers and services

+++ RecipeSocialMediaAPI/Handlers/Users/Commands/RemoveUserCommand.cs
@@ -2,15 +2,16 @@
 using RecipeSocialMediaAPI.DAL;
 using RecipeSocialMediaAPI.DAL.Documents;
 using RecipeSocialMediaAPI.DAL.Repositories;
+using RecipeSocialMediaAPI.Endpoints;
 using RecipeSocialMediaAPI.Handlers.UserTokens.Notifications;
 using RecipeSocialMediaAPI.Services;
 using RecipeSocialMediaAPI.Utilities;
 
 namespace RecipeSocialMediaAPI.Handlers.Users.Commands
 {
-    public record RemoveUserCommand(string Token) : IRequest<bool>;
+    public record RemoveUserCommand(string Token) : IRequest;
 
-    public class RemoveUserHandler : IRequestHandler<RemoveUserCommand, bool>
+    public class RemoveUserHandler : IRequestHandler<RemoveUserCommand>
     {
         private readonly IPublisher _publisher;
         private readonly IUserTokenService _userTokenService;
@@ -23,12 +24,24 @@ namespace RecipeSocialMediaAPI.Handlers.Users.Commands
             _userCollection = factory.GetCollection<UserDocument>(new UserRepository(), config);
         }
 
-        public async Task<bool> Handle(RemoveUserCommand request, CancellationToken cancellationToken)
+        public async Task Handle(RemoveUserCommand request, CancellationToken cancellationToken)
         {
+            if (!_userTokenService.CheckValidToken(request.Token))
+            {
+                throw new InvalidTokenException();
+            }
+
             UserDocument userDoc = _userTokenService.GetUserFromToken(request.Token);
             await _publisher.Publish(new RemoveTokenNotification(request.Token), cancellationToken);
 
-            return await Task.FromResult(_userCollection.Delete(x => x._id == userDoc._id));
+            var successful = await Task.FromResult(_userCollection.Delete(x => x._id == userDoc._id));
+
+            if (!successful)
+            {
+                throw new Exception($"Could not remove user with id {userDoc._id}.");
+            }
+
+            await Task.CompletedTask;
         }
     }
 }
